Feat : change @api_view to follow url rules
- create @api_view PostList, PostDetail so that method is not included in the URL

+++ jobposting/views.py
@@ -6,55 +6,42 @@ from .models import Jobposting
 
 
 # Create your views here.
-@api_view(['GET'])
-def apiOverview(request):
-    api_urls = {
-        'List': '/jobposting-list/',
-        'Detail View': '/jobposting-detail/<str:pk>/',
-        'Create': '/jobposting-create/',
-        'Update': '/jobposting-update/<str:pk>/',
-        'Delete': '/jobposting-delete/<str:pk>/',
-    }
-    return Response(api_urls)
+@api_view(['GET', 'POST'])
+def PostList(request):
+    if request.method == 'GET':
+        jobpostings = Jobposting.objects.all()
+        serializer = JobpostingSerializer(jobpostings, many=True)
+        return Response(serializer.data)
+
+    elif request.method == 'POST':
+        serializer = JobpostingSerializer(data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=201)
+        return Response(serializer.errors, status=404)
+
+@api_view(['GET', 'PUT', 'DELETE'])
+def PostDetail(request, pk):
+    try:
+        jobpostings = Jobposting.objects.get(id=pk)
+    except jobpostings.DoesNotExist:
+        return Response(status=status.HTTP_404_NOT_FOUND)
 
 
-@api_view(['GET'])
-def jobpostingList(request):
-    jobpostings = Jobposting.objects.all()
-    serializer = JobpostingSerializer(jobpostings, many=True)
-    return Response(serializer.data)
+    if request.method == 'GET':
+        serializer = JobpostingSerializer(jobpostings)
+        return Response(serializer.data)
 
 
-@api_view(['GET'])
-def jobpostingDetail(request, pk):
-    jobpostings = Jobposting.objects.get(id=pk)
-    serializer = JobpostingSerializer(jobpostings, many=False)
-    return Response(serializer.data)
+    elif request.method == 'PUT':
+        serializer = JobpostingSerializer(jobpostings, data=request.data)
 
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-@api_view(['POST'])
-def jobpostingCreate(request):
-    serializer = JobpostingSerializer(data=request.data)
-    if serializer.is_valid():
-        serializer.save()
-    return Response(serializer.data)
-
-
-@api_view(['POST'])
-def jobpostingUpdate(request, pk):
-    jobposting = Jobposting.objects.get(id=pk)
-    serializer = JobpostingSerializer(instance=jobposting, data=request.data)
-
-    if serializer.is_valid():
-        serializer.save()
-
-    return Response(serializer.data)
-
-
-@api_view(['DELETE'])
-def jobpostingDelete(request, pk):
-    jobposting = Jobposting.objects.get(id=pk)
-    jobposting.delete()
-
-    return Response('Item succsesfully delete!')
-
+    elif request.method == 'DELETE':
+        jobposting = Jobposting.objects.get(id=pk)
+        jobposting.delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
